+++ esp-hal/src/interrupt/mod.rs
@@ -86,22 +86,19 @@ pub mod software;
 
 #[cfg(feature = "rt")]
 #[unsafe(no_mangle)]
-extern "C" fn EspDefaultHandler(_interrupt: crate::peripherals::Interrupt) {
-    panic!("Unhandled interrupt: {:?}", _interrupt);
+extern "C" fn EspDefaultHandler() {
+    panic!("Unhandled interrupt");
 }
 
 /// Default (unhandled) interrupt handler
 pub const DEFAULT_INTERRUPT_HANDLER: InterruptHandler = InterruptHandler::new(
     {
         unsafe extern "C" {
-            fn EspDefaultHandler(_interrupt: crate::peripherals::Interrupt);
+            fn EspDefaultHandler();
         }
 
         unsafe {
-            core::mem::transmute::<
-                unsafe extern "C" fn(crate::peripherals::Interrupt),
-                extern "C" fn(),
-            >(EspDefaultHandler)
+            core::mem::transmute::<unsafe extern "C" fn(), extern "C" fn()>(EspDefaultHandler)
         }
     },
     Priority::min(),